+++ libnetwork/iptables/firewalld.go
@@ -1,4 +1,4 @@
-//go:build linux
+//go:build linux && !no_systemd
 
 package iptables
 
@@ -6,8 +6,10 @@ import (
 	"context"
 	"fmt"
 	"strings"
+	"time"
 
 	"github.com/containerd/containerd/log"
+	"github.com/docker/docker/pkg/rootless"
 	dbus "github.com/godbus/dbus/v5"
 )
 
@@ -304,4 +306,34 @@ func contains(list []string, val string) bool {
 		}
 	}
 	return false
+}
+
+func initFirewalld() {
+	// When running with RootlessKit, firewalld is running as the root outside our network namespace
+	// https://github.com/moby/moby/issues/43781
+	if rootless.RunningWithRootlessKit() {
+		log.G(context.TODO()).Info("skipping firewalld management for rootless mode")
+		return
+	}
+	if err := firewalldInit(); err != nil {
+		log.G(context.TODO()).WithError(err).Debugf("unable to initialize firewalld; using raw iptables instead")
+	}
+}
+
+// Raw calls 'iptables' system command, passing supplied arguments.
+func (iptable IPTable) Raw(args ...string) ([]byte, error) {
+	if firewalldRunning {
+		// select correct IP version for firewalld
+		ipv := Iptables
+		if iptable.ipVersion == IPv6 {
+			ipv = IP6Tables
+		}
+
+		startTime := time.Now()
+		output, err := Passthrough(ipv, args...)
+		if err == nil || !strings.Contains(err.Error(), "was not provided by any .service files") {
+			return filterOutput(startTime, output, args...), err
+		}
+	}
+	return iptable.raw(args...)
 }
